requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
pillow
python-multipart